+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Versioning](https://semver.org/spec/v2.0.0.html) for `libnopegl`.
   relative anchor for scale and rotate transforms
 - `DrawMask` node to facilitate alpha masking with textures
 - `Text.wrap` parameter to enable text wrapping inside of `Text.box`
+- `DrawTexture.wrap` parameter to select texture wrap behaviour
 
 ### Changed
 - `Text.font_files` text-based parameter is replaced with `Text.font_faces` node